Add management command to publish draft versions that have passed their 'publish_on' time
